# bug/#516 - fixed Marker.hitTest with rotation and optimized drawing