--- a/OpenStreetMapViewer/src/main/java/org/osmdroid/samplefragments/data/SampleMarker.java
+++ b/OpenStreetMapViewer/src/main/java/org/osmdroid/samplefragments/data/SampleMarker.java
@@ -1,13 +1,17 @@
 package org.osmdroid.samplefragments.data;
 
+import android.graphics.drawable.Drawable;
 import android.widget.Toast;
 
 import org.osmdroid.R;
 import org.osmdroid.samplefragments.BaseSampleFragment;
+import org.osmdroid.util.BoundingBox;
 import org.osmdroid.util.GeoPoint;
 import org.osmdroid.views.MapView;
 import org.osmdroid.views.overlay.Marker;
-import org.osmdroid.views.overlay.infowindow.MarkerInfoWindow;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * An example on using osmbonuspack's Marker class by following the tutorial at
@@ -29,21 +33,26 @@ public class SampleMarker extends BaseSampleFragment {
     public void addOverlays(){
         super.addOverlays();
 
+        final List<GeoPoint> points = new ArrayList<>();
+        final Drawable drawable = getResources().getDrawable(R.drawable.icon);
+
         GeoPoint startPoint = new GeoPoint(38.8977, -77.0365);  //white house
+        points.add(startPoint);
         Marker startMarker = new Marker(mMapView);
         startMarker.setPosition(startPoint);
         startMarker.setAnchor(Marker.ANCHOR_CENTER, Marker.ANCHOR_BOTTOM);
-        startMarker.setIcon(getResources().getDrawable(R.drawable.icon));
+        startMarker.setIcon(drawable);
         startMarker.setTitle("White House");
         startMarker.setSnippet("The White House is the official residence and principal workplace of the President of the United States.");
         startMarker.setSubDescription("1600 Pennsylvania Ave NW, Washington, DC 20500");
         mMapView.getOverlays().add(startMarker);
 
         startPoint = new GeoPoint(38.8719, -77.0563);
+        points.add(startPoint);
         startMarker = new Marker(mMapView);
         startMarker.setPosition(startPoint);
         startMarker.setAnchor(Marker.ANCHOR_CENTER, Marker.ANCHOR_BOTTOM);
-        startMarker.setIcon(getResources().getDrawable(R.drawable.icon));
+        startMarker.setIcon(drawable);
         startMarker.setTitle("Pentagon");
         startMarker.setSnippet("The Pentagon.");
         startMarker.setSubDescription("The Pentagon is the headquarters of the United States Department of Defense.");
@@ -59,13 +68,15 @@ public class SampleMarker extends BaseSampleFragment {
 
 
         startPoint = new GeoPoint(38.8895, -77.0353);
+        points.add(startPoint);
         startMarker = new Marker(mMapView);
         startMarker.setPosition(startPoint);
         startMarker.setAnchor(Marker.ANCHOR_CENTER, Marker.ANCHOR_BOTTOM);
-        startMarker.setIcon(getResources().getDrawable(R.drawable.icon));
+        startMarker.setIcon(drawable);
         startMarker.setTitle("Washington Monument");
         startMarker.setSnippet("Washington Monument.");
         startMarker.setSubDescription("Washington Monument.");
+        startMarker.setRotation(45); // for demo purposes
         startMarker.setOnMarkerClickListener(new Marker.OnMarkerClickListener() {
             @Override
             public boolean onMarkerClick(Marker marker, MapView mapView) {
@@ -78,6 +89,12 @@ public class SampleMarker extends BaseSampleFragment {
         mMapView.getOverlays().add(startMarker);
 
 
-        mMapView.invalidate();
+        final BoundingBox boundingBox = BoundingBox.fromGeoPoints(points);
+        mMapView.post(new Runnable() {
+            @Override
+            public void run() {
+                mMapView.zoomToBoundingBox(boundingBox, false, drawable.getIntrinsicWidth());
+            }
+        });
     }
 }