In the process of fixing the initial issue, I optimized the code for Marker drawing, with the following results:
* drawing 5050 markers, all visible - with the latest optim: 130ms, with the old version: 204ms (55%)
* drawing 5050 markers, all visible - with the latest optim: 130ms, without the `BitmapDrawable optim`: 155ms (15%)
* drawing 5050 markers, all visible - with the latest optim: 130ms, without the `rotation==0` optim: 165ms (25%)
* drawing 5050 markers, only one actually visible - with the latest optim: 64ms, with the old version: 200ms (210%)

Impacted classes:
* `DefaultOverlayManager`: unrelated duplicate code fix in method `onDraw`
* `Marker`: added display helper members `mDisplayed`, `mRect` and `mOrientedMarkerRect`; created optimized display method `drawAt`; impacted method `hitTest`; gently refactored
* `SampleMarker`: added a rotation out of nowhere for demo purposes; added a zoom to bounding box feature; gently refactored
* `TileSystem`: added a version of method `getRandomLatitude`
* `SampleMilitaryIconsMarker`: gently refactored around `Random`